refactor: :recycle: move mantine spotlight to own component
