+++ src/interfaces/Spotlight.interface.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react'
+
+export interface ISpotlightProps {
+  children: ReactNode
+}
