+++ src/pages/_app.tsx
@@ -3,21 +3,18 @@ import { useState } from 'react'
 import { AppProps } from 'next/app'
 import { getCookie, setCookies } from 'cookies-next'
 import Head from 'next/head'
-import { useRouter } from 'next/router'
 import {
   MantineProvider,
   ColorScheme,
   ColorSchemeProvider,
 } from '@mantine/core'
-import { SpotlightAction, SpotlightProvider } from '@mantine/spotlight'
 import { NotificationsProvider } from '@mantine/notifications'
+import Spotlight from '../components/Spotlight/Spotlight'
 import GlobalStyles from '../components/GlobalStyles/GlobalStyles'
-import { SpotlightContent } from '../helpers/spotlight'
 
 export default function App(props: AppProps & { colorScheme: ColorScheme }) {
   const { Component, pageProps } = props
   const [colorScheme, setColorScheme] = useState<ColorScheme>(props.colorScheme)
-  const router = useRouter()
 
   const toggleColorScheme = (value?: ColorScheme) => {
     const nextColorScheme = value || (colorScheme === 'dark' ? 'light' : 'dark')
@@ -26,15 +23,6 @@ export default function App(props: AppProps & { colorScheme: ColorScheme }) {
       maxAge: 60 * 60 * 24 * 30,
     })
   }
-
-  const actions: SpotlightAction[] = SpotlightContent?.map(
-    (item: { title: string; url: string }) => ({
-      title: item.title,
-      onTrigger: item.url
-        ? () => router.push(`/${item.url}`)
-        : () => toggleColorScheme(),
-    })
-  )
 
   return (
     <>
@@ -56,11 +44,11 @@ export default function App(props: AppProps & { colorScheme: ColorScheme }) {
           withGlobalStyles
           withNormalizeCSS
         >
-          <SpotlightProvider actions={actions}>
-            <NotificationsProvider>
+          <NotificationsProvider>
+            <Spotlight>
               <Component {...pageProps} />
-            </NotificationsProvider>
-          </SpotlightProvider>
+            </Spotlight>
+          </NotificationsProvider>
         </MantineProvider>
       </ColorSchemeProvider>
     </>
